% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotation_features}
\alias{annotation_features}
\title{Get Features of Annotations}
\usage{
annotation_features(x, pAttribute = c("word", "pos"), annotation)
}
\arguments{
\item{x}{XXX}

\item{pAttribute}{XXX}

\item{annotation}{XXX}
}
\description{
Get Features of Annotations
}
