% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotations_inspect}
\alias{annotations_inspect}
\title{Inspect Annotations Present in a Partition}
\usage{
annotations_inspect(x, annotation = NULL)
}
\arguments{
\item{x}{a partition object}

\item{annotation}{number of an annotation that will be looked up in values for s-attribute}
}
\description{
Inspect Annotations Present in a Partition
}
\examples{
\dontrun{
P <- partition(
  "GERMAPARL",
  lp = 17, session = 175, speaker = "Fritz Rudolf K\\u00F6rper",
  xml = "nested"
)
annotations_inspect(P, annotation = "16")
}
}
