% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\docType{data}
\name{GermaParl}
\alias{GermaParl}
\alias{germaparl_by_lp}
\alias{germaparl_by_year}
\title{GermaParl Corpus Class.}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
GermaParl
}
\description{
The class inherits from the \code{Corpus} class defined in the 
package 'polmineR'. So far, it only adds a method 'summary'.
}
\section{Methods}{

\describe{
  \item{\code{summary(sAttribute)}}{Get number of tokens in subcorpora
  that are prepared on the basis of the s-attribute provided.}
}
}

\examples{
\dontrun{
  germaparl_download_corpus()
  germaparl_download_lda(k = 250L)
  germaparl_add_s_attribute_speech()
  germaparl_encode_lda_topics(k = 250L, n = 5L)
  germaparl_encode_cap_annotations()
  germaparl_build_dtm()
}
}
\keyword{datasets}
