% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.R
\name{germaparl_add_p_attribute_stem}
\alias{germaparl_add_p_attribute_stem}
\title{Add Positional Attribute with Word Stems to GermaParl}
\usage{
germaparl_add_p_attribute_stem(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, whether to output messages.}
}
\value{
Invisibly \code{NULL} is returned.
}
\description{
Add Positional Attribute with Word Stems to GermaParl
}
