% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\name{germaparl_build_dtm}
\alias{germaparl_build_dtm}
\title{Generate Document-Term-Matrix}
\usage{
germaparl_build_dtm(pAttribute = "word", store = TRUE)
}
\arguments{
\item{pAttribute}{The p-attribute for wich to generate the matrix.}

\item{store}{Logical, if \code{TRUE} (default), the dtm will be stored in the package.}
}
\description{
Generate a document-term-matrix from the GermaParl corpus. Speeches will
be treated as documents. The function requires the annotation of speeches
to be present, see \code{germaparl_add_s_attribute_speech}.
}
