% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\docType{data}
\name{germaparl_by_year}
\alias{germaparl_by_year}
\title{Table with information on GermaParl by year}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 22 rows and 6 columns.}
\usage{
germaparl_by_year
}
\description{
A dataset with information on the corpus on a year-by-year basis is included
in the package to be included in the data report of the package vignette. The 
code used to generate the data is reported in the examples section.
}
\details{
Note that the table is based on v1.0.5 of the corpus. The script to reproduce
the table can be found in the file
\url{https://github.com/PolMine/GermaParl/blob/master/data-raw/stats_for_vignette.R}{data-raw/stats_for_vignette.R}.
}
\keyword{datasets}
