% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{germaparl_encode_cap_annotations}
\alias{germaparl_encode_cap_annotations}
\title{Write CAP Annotations to GermaParl}
\usage{

  germaparl_encode_cap_annotations(dir = "/Lab/gitlab/plprbttxt_annotations/2016_09_21")
}
\arguments{
\item{dir}{directory with CAP annotations}
}
\description{
Write CAP Annotations to GermaParl
}
