% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{germaparl_encode_lda_topics}
\alias{germaparl_encode_lda_topics}
\title{Write s-attribute with top topicmodels to corpus}
\usage{
germaparl_encode_lda_topics(k = 200, n = 5)
}
\arguments{
\item{k}{number of topics of the topicmodel to load}

\item{n}{number of topics to write to corpus}
}
\description{
Add a new s-attributes 'topics' to GermaParl corpus with topicmodel 
for \code{k} topics. The \code{n} topics for speeches will be written to 
the corpus.
}
\examples{
\dontrun{
germaparl_encode_lda_topics(k = 250, n = 3)
}
}
