% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\name{germaparl_get_doi}
\alias{germaparl_get_doi}
\title{Get DOI of corpus}
\usage{
germaparl_get_doi(registry_dir = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{registry_dir}{Path to the registry directory.}
}
\value{
If the DOI is declared in the registry file, a length-one
  \code{character} vector with it is returned. If the corpus has not yet been
  installed, \code{NULL} is returned and a warning will be issued.
}
\description{
Get DOI of corpus
}
\details{
Auxiliary function that extracts the DOI (Document Object
  Identifier) from the registry file of the GERMAPARL corpus.
}
\examples{
if (germaparl_is_installed()){
  germaparl_get_doi() # get 'document object identifier' (DOI) of GERMAPARL corpus
}
}
