% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{germaparl_get_speeches_for_topic}
\alias{germaparl_get_speeches_for_topic}
\title{Get Bundle with Speeches for Topic.}
\usage{
germaparl_get_speeches_for_topic(n)
}
\arguments{
\item{n}{topic number}
}
\description{
Get Bundle with Speeches for Topic.
}
\examples{
\dontrun{
k <- 250
model <- germaparl_load_topicmodel(k = k)
T <- topicmodels::terms(model, k)
topic <- 181 # topic nuclear energy
topic <- 133 # foreigners law
topic <- 65 # right-wing extremism

PB <- germaparl_get_speeches_for_topic(topic)
for (i in 1L:length(PB)) {
  fulltext <- read(PB[[i]])
  fulltext <- highlight(fulltext, list(yellow = T[1:100,topic]))
  print(fulltext)
  if (readline() == "q") stop()
}
}
}
