% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\docType{data}
\name{germaparl_lda_tuning}
\alias{germaparl_lda_tuning}
\title{LDA Tuning Results}
\format{
An object of class \code{data.frame} with 10 rows and 5 columns.
}
\usage{
germaparl_lda_tuning
}
\description{
The R package \href{https://CRAN.R-project.org/package=ldatuning}{ldatuning}
has been used to get guidance on the optimal number of topics when fitting an
LDA topic model on the GermaParl corpus. Using around 250 topics is a good
choice. The data object \code{germaparl_lda_tuning} reports the different metrics of the
ldatuning package.
}
\keyword{datasets}
