% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{germaparl_load_topicmodel}
\alias{germaparl_load_topicmodel}
\title{Load Topicmodel}
\usage{
germaparl_load_topicmodel(k, verbose = TRUE)
}
\arguments{
\item{k}{number of topics}

\item{verbose}{logical}
}
\description{
Load Topicmodel
}
