% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{germaparl_regdir}
\alias{germaparl_regdir}
\title{Get the registry directory in the GermaParl package.}
\usage{
germaparl_regdir()
}
\value{
The registry directory within the GermaParl package.
}
\description{
Utility function that returns the registry directory within the GermaParl package.
}
