% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{germaparl_search_speeches}
\alias{germaparl_search_speeches}
\title{Query GermaParl}
\usage{
germaparl_search_speeches(cnt, min_size = 250)
}
\arguments{
\item{cnt}{XXX}

\item{min_size}{XXX}
}
\description{
Query GermaParl
}
\examples{
\dontrun{
P <- partition("GERMAPARL", cap = "^.*\\\\|8-01\\\\|.*$", regex = TRUE)
C <- count(P, pAttribute = c("word", "pos"))
CNT <- as(C, "count")
matches <- query(cnt = CNT, min_size = 250)
PB <- partitionBundle("GERMAPARL", sAttribute = "speech", values = names(matches)[1:20])
}
}
