% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{germaparl_download_corpus}
\alias{germaparl_download_corpus}
\title{Download GermaParl from Web Directory.}
\usage{
germaparl_download_corpus(tarball = "germaparl.tar.gz",
  dir = "http://polmine.sowi.uni-due.de/corpora/cwb/germaparl")
}
\arguments{
\item{tarball}{Name of the tarball.}

\item{dir}{directory where the tarball resides, either the URL of a webspace,
or a local directory}
}
\description{
\code{germaparl_download_corpus} will get a tarball with the indexed corpus
from a designated web space or a local directory and install the corpus into
the GermaParl package.
}
