% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_germaparl}
\alias{use_germaparl}
\title{Use the GermaParl Corpus.}
\usage{
use_germaparl()
}
\value{
Invisibly, the registry directory in the GermaParl package is returned.
}
\description{
Utility function to ensure that registry directory points to registry in
GermaParl package.
}
