% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\docType{package}
\name{GermaParl-package}
\alias{GermaParl-package}
\alias{GermaParl}
\title{GermaParl R Data Package.}
\arguments{
\item{registry_dir}{Path to the system registry directory.}
}
\description{
The package offers a convenient dissemination mechanism for the 
GermaParl corpus that has been linguistically annotated and indexed (data format
of the Corpus Workbench / CWB). After installation, only a small sample corpus
will be included in the package. Use the \code{germaparl_download_corpus} function
to download the full corpus. The package offers further functionality to amend 
the corpus.
}
\details{
Note that the R package and the corpus are two different kinds of research
data: The package offers a mechanism to ship, easily install and augment the data.
The indexed corpus is the actual data. Package and corpus have different version
numbers and should be quoted in combination in publications. We recommend to follow 
the instructions you see when calling \code{citation(package = "GermaParl")}.
}
\references{
Blaette, Andreas (2018): "Using Data Packages to Ship Annotated
  Corpora of Parliamentary Protocols: The GermaParl R Package". ISBN
  979-10-95546-02-3.
}
\author{
Andreas Blaette \email{andreas.blaette@uni-due.de}
}
\keyword{package}
