% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\docType{package}
\name{GermaParl-package}
\alias{GermaParl-package}
\alias{GermaParl}
\title{GermaParl R package.}
\description{
GermaParl R package.
}
\examples{
\dontrun{
library(polmineR)
use("GermaParl")
corpus() # will include GERMAPARLMINI, sample corpus included in pkg
germaparl_download_corpus() # takes time!
use("GermaParl")
corpus() # will include GERMAPARL, full corpus
}
}
\author{
Andreas Blaette (andreas.blaette@uni-due.de)
}
\keyword{package}
