% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\name{GermaParl}
\alias{GermaParl}
\alias{germaparl_by_lp}
\alias{germaparl_by_year}
\title{GermaParl Corpus Class.}
\description{
The class inherits from the \code{Corpus} class defined in the 
package 'polmineR'. So far, it only adds a method 'summary'.
}
\section{Methods}{

\describe{
  \item{\code{summary(sAttribute)}}{Get number of tokens in subcorpora
  that are prepared on the basis of the s-attribute provided.}
}
}

\examples{
\dontrun{
  germaparl_download_corpus()
  germaparl_download_lda(k = 250L)
  germaparl_add_s_attribute_speech()
  germaparl_encode_lda_topics(k = 250L, n = 5L)
  germaparl_encode_cap_annotations()
  germaparl_build_dtm()
}
}
\section{Super class}{
\code{\link[polmineR:Corpus]{polmineR::Corpus}} -> \code{GermaParl}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GermaParl$new()}}
\item \href{#method-summary}{\code{GermaParl$summary()}}
\item \href{#method-clone}{\code{GermaParl$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermaParl$new(corpus = "GERMAPARL", pAttribute = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermaParl$summary(sAttribute)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermaParl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
