% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\docType{data}
\name{germaparl_by_year}
\alias{germaparl_by_year}
\title{Table with information on GermaParl by year}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 22 rows and 6 columns.}
\usage{
germaparl_by_year
}
\description{
Table with information on GermaParl by year
}
\examples{
\dontrun{
dts <- lapply(
 13:17,
 function(lp){
  print(lp)
  P <- partition("GERMAPARL", lp = lp)
  dates <- as.Date(s_attributes(P, "date"))
  dt <- count(P, p_attribute = "lemma")
  unknown <- round(sum(dt[grepl("#unknown#", dt[["lemma"]])][["count"]]) / size(P), digits = 3)
  
  data.table(
    lp = lp,
    protocols = length(unique(s_attributes(P, "session"))),
    first = min(dates, na.rm = TRUE),
    last = min(dates, na.rm = TRUE),
    size = size(P),
    unknown = unknown
  )
}
)
germaparl_by_lp <- rbindlist(dts)
}
}
\keyword{datasets}
