% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\name{germaparl_get_doi}
\alias{germaparl_get_doi}
\title{Get DOI of corpus}
\usage{
germaparl_get_doi(registry_dir = Sys.getenv("CORPUS_REGISTRY"), sample = FALSE)
}
\arguments{
\item{registry_dir}{Path to the registry directory.}

\item{sample}{A \code{logical} value, if \code{FALSE} (default), use
GERMAPARL, if \code{TRUE}, use GERMAPARLSAMPLE.}
}
\value{
If the DOI is declared in the registry file, a length-one
  \code{character} vector with it is returned. If the corpus has not yet been
  installed, \code{NULL} is returned and a warning will be issued.
}
\description{
Auxiliary function that extracts the DOI (Document Object Identifier) from
the registry file of the GERMAPARL corpus.
}
\seealso{
See the examples section of the overview documentation of the
  \link{GermaParl} package for an example.
}
