% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\name{germaparl_is_installed}
\alias{germaparl_is_installed}
\title{Get installation status of GERMAPARL}
\usage{
germaparl_is_installed(registry_dir = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{registry_dir}{Path to the registry directory.
Auxiliary function to detect whether GERMAPARL is installed or not.}
}
\value{
\code{TRUE} if the corpus has been installed, and \code{FALSE} if not.
}
\description{
Get installation status of GERMAPARL
}
\examples{
germaparl_is_installed() # to check whether GERMAPARL has been downloaded
}
