% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{germaparl_download_corpus}
\alias{germaparl_download_corpus}
\title{Download GermaParl from Web Directory.}
\usage{
germaparl_download_corpus(
  doi = "https://doi.org/10.5281/zenodo.3735141",
  quiet = FALSE
)
}
\arguments{
\item{doi}{The DOI (Digital Object Identifier) of the GermaParl tarball at zenodo.}

\item{quiet}{Whether to output messages.}
}
\description{
\code{germaparl_download_corpus} will get a tarball with the indexed corpus
from a designated web space or a local directory and install the corpus into
the GermaParl package.
}
