% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GermaParl.R
\docType{data}
\name{lda_tuning}
\alias{lda_tuning}
\title{LDA Tuning Results}
\format{An object of class \code{data.frame} with 10 rows and 5 columns.}
\usage{
lda_tuning
}
\description{
LDA Tuning Results
}
\keyword{datasets}
